+++ src/script/components/VerificationBadge/index.ts
